backend/requirements.txt
asgiref==3.11.0
channels==4.3.2
daphne>=4.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.15.0
pytest==7.4.0
pytest-django==4.5.2
pytest-asyncio==0.22.0
